==> 11_Model.ipynb <==
import pandas as pd  # таблицы
import joblib

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.multioutput import MultiOutputClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, classification_report

dir = "D:\Develop\VS\HSE\Pump_Dump\\"  # папка проекта

main_df = pd.read_csv(dir + "Answers\\news_quotes.csv", encoding="utf-8", sep="|", index_col=0)
main_df.head()

imoex_df = pd.read_csv(dir + "Answers\\imoex_changes.csv", encoding="utf-8", sep="|", index_col=0)
imoex_df.head()

# проводим сентимент-анализ
# from textblob import TextBlob
# from textblob.sentiments import NaiveBayesAnalyzer # NLTK model trained on a movie reviews corpus

# def sentiment(text):
#     sent = TextBlob(text, analyzer=NaiveBayesAnalyzer())
#     return sent.sentiment

# main_df["Mood"] = main_df["WORDS_lemma"].apply(lambda x: sentiment(x))

# рассчитываем изменения цен до и после и категорируем их
# диапазоны изменения для категорирования взяты экспертно
borders = [-10, -0.6, 0.6, 10] # 0 - negative, 1 - neutral, 2 - positive

imoex_df["Change_B"] = round((imoex_df["Price"] - imoex_df["Before"])/imoex_df["Price"]*100,2)
imoex_df["Change_A"] = round((imoex_df["After"] - imoex_df["Price"])/imoex_df["Price"]*100,2)

imoex_df["Mood_B"] = pd.cut(imoex_df['Change_B'], borders).astype("category").cat.codes
imoex_df["Mood"] = pd.cut(imoex_df['Change'], borders).astype("category").cat.codes
imoex_df["Mood_A"] = pd.cut(imoex_df['Change_A'], borders).astype("category").cat.codes

imoex_df = pd.get_dummies(imoex_df, columns=["Mood"], dtype=int)
imoex_df = pd.get_dummies(imoex_df, columns=["Mood_B"], dtype=int)
imoex_df = pd.get_dummies(imoex_df, columns=["Mood_A"], dtype=int)
imoex_df.head()


# если полный файл был сформирован ранее, то используем его и новые данные не подгружаем IMOEX. При изменении границ категорирования файл надо обновить отдельно

# объединяем новости и цены ИНДЕКСА
# main_df = pd.merge(
#     main_df, imoex_df, left_on=["NEWS_DATE"], right_on=["Date"], how="left"
# )

# main_df.head()

# формируем фрейм для обучения модели
model_df = main_df[["WORDS_lemma", "Mood_0","Mood_1","Mood_2","Mood_B_0","Mood_B_1","Mood_B_2","Mood_A_0","Mood_A_1","Mood_A_2"]]
model_df.dropna(inplace = True)
model_df.head()


# формируем обучающую и тестовые выборки
X = model_df["WORDS_lemma"].values.astype('U')
Y = model_df[model_df.columns[1:10]]

X_train, X_test, y_train, y_test = train_test_split(X, Y, random_state=42, test_size=0.2)
X_train.shape, y_train.shape, X_test.shape, y_test.shape


# проводим векторизацию. Закомментированы альтернативные векторайзеры

# vectorizer = CountVectorizer(ngram_range=(1, 3))
# X_train = vectorizer.fit_transform(X_train)
# X_test = vectorizer.transform(X_test)

vectorizer = TfidfVectorizer()
X_train = vectorizer.fit_transform(X_train)
X_test = vectorizer.transform(X_test)

# Создание многозадачной модели с использованием случайного леса
multi_task_model = MultiOutputClassifier(RandomForestClassifier())

# Обучение модели на данных
multi_task_model.fit(X_train, y_train)

# Сохраняем обученную модель. !!! имя модели соответствует различным векторайзерам
# joblib.dumpa(multi_task_model, dir + "Answers\\model_classif_TFIDF.pkl") 

# проводим предсказание и смотрим на качество работы модели

y_predict = multi_task_model.predict(X_test)
print("TFIDF_accuracy: ", accuracy_score(y_predict, y_test))
print(classification_report(y_test, y_predict))

# сохраняем собранный фрейм на диск
# main_df.to_csv(dir + "Answers\\news_quotes.csv", encoding="utf-8", sep="|")

# Загружаем обученную модель. !!! имя модели соответсвует различным векторайзерам
# multi_task_model = joblib.load("Answers\\model_classif_TFIDF.pkl")
