==> 12_Summary.ipynb <==
import pandas as pd  # таблицы

from transformers import MBartTokenizer, MBartForConditionalGeneration
from itertools import combinations
import torch

from nltk.tokenize import sent_tokenize, RegexpTokenizer
from nltk.stem.snowball import RussianStemmer
import networkx as nx


dir = "D:\Develop\VS\HSE\Pump_Dump\\"  # папка проекта
main_df = pd.read_csv(dir + "Answers\\news_quotes.csv", encoding="utf-8", sep="|", index_col=0)
main_df.head()
sum_df = main_df[(main_df["TICKERS"].str.len()>2) & (main_df["TICKERS"].str.len()<5) & (main_df["TICKERS"]=="lkoh")][["NEWS_TEXT", "TICKERS"]][:20]
sum_df


model_name = "IlyaGusev/mbart_ru_sum_gazeta"
tokenizer = MBartTokenizer.from_pretrained(model_name)

model = MBartForConditionalGeneration.from_pretrained(model_name)

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model.to(device)

def gen_summary(article_text):
    print(article_text)
    input_ids = tokenizer(
        [article_text],
        max_length=600,
        truncation=True,
        return_tensors="pt",
    )["input_ids"].to(device)
    output_ids = model.generate(
        input_ids=input_ids,
        no_repeat_ngram_size=4
    )[0]
    summary = tokenizer.decode(output_ids, skip_special_tokens=True)
    return summary

sum_df["Summary_Bart"] = sum_df["NEWS_TEXT"].apply(gen_summary)

# Обрабатываем текст на основе PageRank

def similarity(s1, s2):
    if not len(s1) or not len(s2):
        return 0.0
    return len(s1.intersection(s2))/(1.0 * (len(s1) + len(s2)))

# Выдает список предложений отсортированных по значимости
def textrank(text):
    sentences = sent_tokenize(text)
    tokenizer = RegexpTokenizer(r'\w+')
    lmtzr = RussianStemmer()
    words = [set(lmtzr.stem(word) for word in tokenizer.tokenize(sentence.lower()))
             for sentence in sentences] 	 
    pairs = combinations(range(len(sentences)), 2)
    scores = [(i, j, similarity(words[i], words[j])) for i, j in pairs]
    scores = filter(lambda x: x[2], scores)
    g = nx.Graph()
    g.add_weighted_edges_from(scores)
    pr = nx.pagerank(g)
    return sorted(((i, pr[i], s) for i, s in enumerate(sentences) if i in pr), key=lambda x: pr[x[0]], reverse=True)

# Сокращает текст до нескольких наиболее важных предложений
def sumextract(text, n=5):
    tr = textrank(text)
    top_n = sorted(tr[:n])
    return ' '.join(x[2] for x in top_n)

sum_df["Summary_PageRank"] = sum_df["NEWS_TEXT"].apply(sumextract)

sum_df

from datasets import load_metric
metric = load_metric("rouge", trust_remote_code=True)

def calc_rouge_scores(candidates, references):
    result = metric.compute(predictions=candidates, references=references, use_stemmer=True)
    result = {key: round(value.mid.fmeasure * 100, 1) for key, value in result.items()}
    return result


print("Bart - ",calc_rouge_scores(sum_df["NEWS_TEXT"], sum_df["Summary_Bart"]))
print("PageRank - ",calc_rouge_scores(sum_df["NEWS_TEXT"], sum_df["Summary_PageRank"]))

sum_df[["NEWS_TEXT","Summary_Bart","Summary_PageRank"]].to_html(dir+"Answers\\compare.html")
# main_df.to_csv(dir + "Answers\\news_quotes.csv", encoding="utf-8", sep="|")
